+++ gamemodes/terrortown/gamemode/shared/lang/de.lua
@@ -124,7 +124,6 @@ L.radar_charging = "Dein Radar lädt immer noch auf!"
 L.xfer_name = "Transfer"
 L.xfer_menutitle = "Credits transferieren"
 L.xfer_send = "Sende einen Credit"
-L.xfer_help = "Du kannst Credits nur an weitere {role} senden."
 
 L.xfer_no_recip = "Der Empfänger ist ungültig, Credit-Transfer abgebrochen."
 L.xfer_no_credits = "Ungenügend Credits für einen Transfer."